refactor[#345] : 기본정보 추가 api 요청 함수에 apiClient 적용
