+++ src/features/auth/api/index.ts
@@ -1,5 +1,5 @@
 export * from "./postLogin.ts";
-export * from "./signUp.ts";
+export * from "./putAddUserInfo.ts";
 export * from "./postAddPetInfo.ts";
 export * from "./putChangeNickname.ts";
 export * from "./postSendCode.ts";
